Committing new files
